clears unused components from Removed OtpScreen and refactors OtpCodeComponent.tsx position
