+++ ts/features/bonus/cgn/components/merchants/discount/CgnOTPCodeComponent.tsx
@@ -9,13 +9,13 @@ import { useIODispatch, useIOSelector } from "../../../../../../store/hooks";
 import { cgnOtpDataSelector } from "../../../store/reducers/otp";
 import { isError, isLoading, isReady } from "../../../../bpd/model/RemoteValue";
 import { cgnGenerateOtp, resetOtpState } from "../../../store/actions/otp";
-import { OtpCodeComponent } from "../../otp/OtpCodeComponent";
 import Eye from "../../../../../../../img/icons/Eye.svg";
 import ActivityIndicator from "../../../../../../components/ui/ActivityIndicator";
 import { Link } from "../../../../../../components/core/typography/Link";
 import { IOStyles } from "../../../../../../components/core/variables/IOStyles";
 import { H3 } from "../../../../../../components/core/typography/H3";
 import { H4 } from "../../../../../../components/core/typography/H4";
+import { OtpCodeComponent } from "./OtpCodeComponent";
 
 const styles = StyleSheet.create({
   row: {
