+++ ts/features/bonus/cgn/components/merchants/discount/OtpCodeComponent.tsx
@@ -2,16 +2,16 @@ import React, { useEffect, useRef, useState } from "react";
 import { Animated, Easing, StyleSheet, View, ViewStyle } from "react-native";
 import { Millisecond, Second } from "italia-ts-commons/lib/units";
 import { fromNullable } from "fp-ts/lib/Option";
-import { Otp } from "../../../../../../definitions/cgn/Otp";
-import { IOColors } from "../../../../../components/core/variables/IOColors";
-import { addEvery } from "../../../../../utils/strings";
-import { BaseTypography } from "../../../../../components/core/typography/BaseTypography";
-import IconFont from "../../../../../components/ui/IconFont";
-import TouchableDefaultOpacity from "../../../../../components/TouchableDefaultOpacity";
-import { clipboardSetStringWithFeedback } from "../../../../../utils/clipboard";
-import { isTestEnv } from "../../../../../utils/environment";
-import I18n from "../../../../../i18n";
-import { H5 } from "../../../../../components/core/typography/H5";
+import { Otp } from "../../../../../../../definitions/cgn/Otp";
+import { IOColors } from "../../../../../../components/core/variables/IOColors";
+import { addEvery } from "../../../../../../utils/strings";
+import { BaseTypography } from "../../../../../../components/core/typography/BaseTypography";
+import IconFont from "../../../../../../components/ui/IconFont";
+import TouchableDefaultOpacity from "../../../../../../components/TouchableDefaultOpacity";
+import { clipboardSetStringWithFeedback } from "../../../../../../utils/clipboard";
+import { isTestEnv } from "../../../../../../utils/environment";
+import I18n from "../../../../../../i18n";
+import { H5 } from "../../../../../../components/core/typography/H5";
 
 type ProgressConfig = {
   startPercentage: number;
